Fix of date in bad format when resending accesses
When resending access or limited access, they are retrieved from
the central DB.
Time columns in MariaDB are retrieved as timedelta type.
If it is converted to string using str() function, something
like 0:27:00 is got. When it is sent to controller in this way,
it causes problems when it is compared with times like 09:23.
(For example: '0:00:00' > '09:31' returns True).
For this reason, it should be formatted and sent in the XX:XX format.
To format it in this way, it should be converted to datetime object
and formatted using strftime() function.
To do it, the timedelta object should be added to the minimum datetime
object, then retrieve the time from it and call the strftime() method.