Enhancement in re-provisioning mechanism
Before this enhancement, when controller reprov endpoint was called,
the server was setting all the CRUDS of this controller to state: TO_ADD,
then it was sending to the controller, the message RRP. The controller
was answering to this message in the same way as when it was asked to
resend an specific CRUD.
The response was entering to "CrudReSndr" thread as the MAC of the
controller an "CrudReSndr" was sending all the CRUDs set to
state: TO_ADD previously.
Although this solution looks smart and simple, it had a problem:
If a controller that is not connected or reachable was asked to
re-provisioning, the CRUDs of this controller was set to state: TO_ADD in
the server DB before sending the RRP message to controller. As the
controller wasn't connected and couldn't receive this message, it wasn't
clearing its DB and it wasn't answering with the response.
Then, when the controller was connecting again, the "CrudReSndr" would
have tryied to send all the CRUDs set to state: TO_ADD, but as the controller
had never cleared the its DB, for sure the operation wouldn't had success.
Now, when controller reprov endpoint was called, the server just send the RRP
message to the controller. When the controller receives that message, it
clears the its DB and after that it answer with RRRP. When the network thread
of the server receives that response, it tells the message receiver thread to
set all the CRUDs of this controller to state: TO_ADD, and the message
receiver thread tells the CRUD resender thread to send all the CURDs of this
controller sending the MAC of this controller in the same way the network
thread sends messages to this thread (putting the MAC of the controller
in "toCrudReSndr" queue)